Enhance --> displayNotificationMessage method
