+++ app/front/static/js/services/cart-service.js
@@ -1,6 +1,7 @@
 import { Helpers } from "../helpers/helpers.js"
 import { cartItemTemplate } from "../components/cart-item.js"
 
+
 export class CartService {
 
     constructor() {
@@ -39,7 +40,7 @@ export class CartService {
 
                             if (!item.no_stock) {
                                 const message = "Producto actualizado exitosamente"
-                                this.toolbox.displayNotificationMessage(message)
+                                this.helpers.displayNotificationMessage(message, "success")
                             }
                         } else {
                             const itemElement = document.createElement("div")
@@ -49,7 +50,7 @@ export class CartService {
                             cartItemsContainer.appendChild(itemElement)
 
                             const message = "Producto agregado exitosamente"
-                            this.toolbox.displayNotificationMessage(message)
+                            this.helpers.displayNotificationMessage(message, "success")
                         }
 
                         const newSubtractButtons = document.querySelectorAll(`[id^="subtract_quantity_${uuid}"]`)
@@ -97,7 +98,7 @@ export class CartService {
                         itemElement.remove()
 
                         const message = "Producto removido exitosamente"
-                        this.toolbox.displayNotificationMessage(message)
+                        this.helpers.displayNotificationMessage(message, "success")
 
                         this.toolbox.updateBubbleCounter(data.cart.items)
                     }
@@ -169,49 +170,6 @@ class Toolbox {
         const bubbleCounter = document.getElementById("bubble_counter")
         const itemCount = Object.keys(items).length
         bubbleCounter.textContent = itemCount
-    }
-
-
-    /**
-    * Display Notification:
-    * @param {string} message Message.
-    * @returns {void} None.
-    */
-    displayNotificationMessage(message) {
-        const messageContainer = document.getElementById("message_container")
-
-        const msg = document.createElement("div")
-        msg.id = "message"
-        msg.classList = "z-10 fixed top-24 flex justify-end w-full p-4 transform translate-x-full transition-transform duration-300 ease-in-out"
-        msg.innerHTML = `
-            <div class="flex flex-row items-center justify-between w-full sm:w-1/2 lg:w-1/3 p-4 bg-green-50 
-                text-md md:text-lg text-green-600 text-center border-l-2 border-green-600 rounded-md">
-                <div>
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" 
-                        stroke-width="1.5" stroke="currentColor" class="size-5 md:size-6">
-                        <path stroke-linecap="round" stroke-linejoin="round" 
-                            d="M9 12.75 11.25 15 15 9.75M21 12a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z" />
-                    </svg>              
-                </div>
-                <div class="flex items-center justify-center w-full">
-                    <div id="message_content" class="mukta-light">
-                        ${message}
-                    </div>
-                </div>
-            </div>
-        `
-        messageContainer.appendChild(msg)
-
-        setTimeout(() => {
-            msg.classList.toggle("translate-x-full")
-            setTimeout(() => {
-                msg.classList.toggle("translate-x-full")
-                setTimeout(() => {
-                    msg.remove()
-                }, 2300)
-            }, 2200)
-        }, 300)
-
     }
 
 }
